Added chat functionailty

+++ app/controllers/api/v1/friend_requests_controller.rb
@@ -5,7 +5,6 @@ module Api
             def send_friend_request
                 friend = User.find(params[:friend_id])
                 friend_request = FriendRequest.new(sent_by: @current_user, sent_to: friend, status: false)
-          
                 if friend_request.save
                   render json: { message: 'Friend request sent successfully.' }
                 else
@@ -13,15 +12,15 @@ module Api
                 end
             end
 
-            def destroy
-              friend = User.find(params[:id])
-            end
-
             def reject_friend_request
               friend = User.find(params[:id])
               friend_request= FriendRequest.find_by(sent_by: friend.id, sent_to:@current_user.id)
-              friend_request.destroy
-              render json: { message: 'Friend request rejected.' }
+              if friend_request
+                friend_request.destroy
+                render json: { message: 'Friend request rejected.' }
+              else
+                render json: { error: 'Friend request not found.' }, status: :not_found
+              end
             end
 
             def received_friend_requests
@@ -34,8 +33,9 @@ module Api
 
             def sent_friend_requests
               current_user= User.find(params[:id])
-              sent_requests= current_user.received_friend_requests
-              recievers_ids=sent_requests..select { |request| request.status == false }.map { |request| request.sent_to_id }
+              sent_requests= current_user.sent_friend_requests
+              recievers_ids=sent_requests.select { |request| request.status == false }.map { |request| request.sent_to_id }
+              render json: recievers_ids
             end
         end
     end
